# documentation/swagger.json
{
  "info": {
    "title": "Baraka API documentation",
    "version": "1.0.0",
    "license": {
      "name": "Sous-traitant Baraka"
    },
    "contact": {
      "email": "abdourahime.mohamadi-halidi@epitech.eu"
    },
    "description": "API REST pour gérer l'application Baraka"
  },
  "basePath": "/api/v1",
    "host": "https://baraka-api.herokuapp.com/",
...
  "paths": {
...
    "/bar/{barname}": {
      "post": {
        "tags": ["Bars"],
        "summary": "/bar/{barname}",
        "consumes": [
          "application/json"
        ],
        "parameters": [{
            "name": "barname",
            "in": "body",
            "required": true,
            "schema": {
              "description": "On renseigne le nom du bar rechercher",
              "type": "string",
              "example": "Wayne's bar"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK return bar(s)",
            "examples": {
              "application/json": {
                "id": "5e16f814e163d4001717d4ab",
                "name" : "Wayne's Bar",
                "description": "Dans une ambiance 'Retro American' est le décor pour boire une bière dans la 'french riviera",
                "coordgps": "15 Rue de la Préfecture, 06300 Nice",
                "averageNotation": 4.5,
                "trends": [
                  "trend","cool","stylié"
                ]
              }
            }
          },
          "422":{
            "description": "Error.",
            "examples": {
              "application/json": {
                "Error" : "Une erreur s'est produite pour trouver votre bar."
              }
            }
          }
        }
      }
    },
...
    "/delete-favorite/{id}": {
      "delete": {
        "tags": ["Bars"],
        "description": "L'utilisateur doit être connecté pour utiliser cette route",
        "summary": "Supprimer un bar de ses favoris.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "barID",
            "in": "body",
            "required": true,
            "schema": {
              "description": "On renseigne le nom du bar",
              "type": "string",
              "example": "5e16f77ce163d4001717d4aa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "retourne OK",
            "examples": {
              "application/json": {
                "success": "OK"
              }
            }
          },
          "422":{
            "description": "Error",
            "examples": {
              "application/json": {
                "Error" : "Ce bar n'existe pas dans vos favoris."
              }
            }
          },
          "422-bis":{
            "description": "Error",
            "examples": {
              "application/json": {
                "Error" : "Une erreur s'est produite pour supprimer ce bar de vos favoris."
              }
            }
          }
        }
      }
    },
    "/my-favorite-bar": {
      "get": {
        "tags": ["Bars"],
        "description": "L'utilisateur doit être connecté pour utiliser cette route",
        "summary": "Liste les bars favoris de l'utilisateur.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "OK retourne un ou des bars.",
            "examples": {
              "application/json": {
                "id": "5e16f77ce163d4001717d4aa",
                "name" : "Chez Tonton",
                "description": "Le QG des étudiants de Toulouse. Il faut dire qu'avec des pressions demi à 2,30 €, des pintes à 4 €, sans oublier le fameux pastis à 2,50 €, les étudiants sont vite conquis",
                "coordgps": "16 pl St Pierre 31000 Toulouse",
                "averageNotation": 4.5,
                "trends": [
                  "trend","cool","stylié"
                ]
              }
            }
          },
          "422":{
            "description": "Error",
            "examples": {
              "application/json": {
                "Error" : "Une Error s'est produite pour lister vos favoris."
              }
            }
          }
        }
      }
    },
...
    "/bar/all-comment/{barID}": {
      "get": {
        "tags": ["Bars"],
        "description": "L'utilisateur doit être connecté pour utiliser cette route",
        "summary": "Lister tous les commentaires.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "barID",
            "in": "body",
            "required": true,
            "schema": {
              "description": "On renseigne l'ID du bar qu'on souhaite lister les commentaires.",
              "type": "string",
              "example": "5e16f77ce163d4001717d4aa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK retourne un ou plusieurs commentaires.",
            "examples": {
              "application/json": {
                "id": "5e16f77ce163d4001717d4aa",
                "comment" : "Une soirée de ouf !",
                "author": "toto",
                "image": "iVBORw0KGgoAAAANSUhEUgAAANcAAADXCAIAAAAGH1PiAAAACXBIWXMAAAsSAAALEgHS3X78AAAgAElEQVR42u2d91dUXbrn..."
              }

            }
          },
          "422":{
            "description": "Error",
            "examples": {
              "application/json": {
                "Error" : "Une Error s'est produite pour lister les commentaires."
              }
            }
          }
        }
      }
    },
...
  },
  "swagger": "2.0"
}